+++ brutils/__init__.py
@@ -31,3 +31,9 @@ from brutils.legal_process import (
     format_processo_juridico,
     remove_symbols as remove_symbols_legal_process,
 )
+from brutils.license_plate import is_valid_license_plate_old_format
+
+from brutils.pis import (
+    is_valid as is_valid_pis,
+    generate as generate_pis,
+)